注解校验完善，修复已知问题

+++ coolGuard-decision/src/main/java/cn/wnhyang/coolGuard/indicator/CountIndicator.java
@@ -28,7 +28,7 @@ public class CountIndicator extends AbstractIndicator {
     @Override
     public Object getResult0(IndicatorContext.IndicatorCtx indicator, RScoredSortedSet<String> set) {
 
-        return (double) set.size();
+        return set.size();
     }
 
     @Override
